Cast to remove warning
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@440 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/WizardConnectionPage.cpp
@@ -261,7 +261,7 @@ LRESULT CWizardConnectionPage::OnWizardNext()
 		Settings.Connection.OutSpeed = nSpeed;		//Cable, SDSL, and the big boys.
 
 	//Set upload limit to 90% of capacity, trimmed to the nearest KB.
-	Settings.Bandwidth.Uploads = Settings.Connection.OutSpeed * 0.9;
+	Settings.Bandwidth.Uploads = (DWORD)( Settings.Connection.OutSpeed * 0.9 );
 	Settings.Bandwidth.Uploads >>= 3;
 	Settings.Bandwidth.Uploads *= 1024;
